% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_bins.R
\name{climate_bins}
\alias{climate_bins}
\title{Climatic bins}
\usage{
climate_bins(interval)
}
\arguments{
\item{interval}{A character string matching an interval available in NAME of \code{GTS2020}.}
}
\value{
A \code{data.frame} of climatic bins for a specified interval.
}
\description{
A function which return climatic bins for a given time interval.
}
\examples{
lat_bins(size = 20)
lat_bins(size = 13, fit = TRUE)
}
