% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{gradient generates a latitudinal temperature gradient
using a generalised logistic function}
\usage{
gradient(x, coeff, sdy)
}
\arguments{
\item{x}{absolute latitudes}

\item{coeff}{coefficients of the generalised logistic function}

\item{sdy}{standard deviation of the points from the curve}
}
\value{
one temperature values for each latitude x, based on coeff and sdy
}
\description{
gradient generates a latitudinal temperature gradient
using a generalised logistic function
}
\examples{

}
