% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_bins.R
\name{climate_bins}
\alias{climate_bins}
\title{Climatic bins}
\usage{
climate_bins(interval)
}
\arguments{
\item{interval}{\code{character or numeric}. Interval name or age available in \code{GTS2020}. If a single interval name is provided, this interval is returned.
If two interval names are provided, these intervals and those existing between are returned. If a single interval age is provided, the age matching this interval is returned.
If two interval ages are provided, the intervals occurring in the range of these ages are returned.}
}
\value{
A \code{data.frame} of climatic bins for a specified interval.
}
\description{
A function which return climatic bins for a given time interval.
}
\examples{
Using a single interval
climate_bins(interval = "Maastrichtian")
Using multiple intervals
climate_bins(interval = c("Albian", "Maastrichtian"))
Using interval midpoint age
climate_bins(interval = 100)
Using interval age range
climate_bins(interval = c(50, 100))
}
